% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CreateCustomTimeSeries}
\alias{CreateCustomTimeSeries}
\title{Create a Custom Time Series}
\usage{
CreateCustomTimeSeries(values, intervals)
}
\arguments{
\item{values}{A numeric vector of values for the time series}

\item{intervals}{A numeric vector of time intervals for the time series. This should be one
item shorter than values. Intervals are set in seconds where 60 would equal 1 minute.}
}
\value{
An xts time series
}
\description{
Create an xts time series using an vector of values and a vector of time intervals.
}

