% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_patterns.R
\name{dropPhase}
\alias{dropPhase}
\title{plasma-Ca}
\usage{
dropPhase(timeseries, index)
}
\arguments{
\item{timeseries}{The time series}

\item{index}{The index of the current point}
}
\description{
Contains a Drop Phase
}

