% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pips.R
\name{GetAdjacentPIPIndex}
\alias{GetAdjacentPIPIndex}
\title{Get one of the Adjacent Perceptually Important Points (PIPs) to the Point Provided}
\usage{
GetAdjacentPIPIndex(index, is.pip, side)
}
\arguments{
\item{index}{The numeric index of the point in is.pip, for which the adjacent pip
will be retrieved}

\item{is.pip}{A logical vector of points of the same length of the time series
in question. TRUE, indicates that the point is a pip.}

\item{side}{May be either "left" or "right". Designated the side on which the adjacent
pip is to be found}
}
\value{
The index of the closes pip on the designated side of the given point.
}
\description{
Retreives a the nearest PIP to the point designated by index on the side designated
by side
}

