% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pips.R
\name{GetPerpDist}
\alias{GetPerpDist}
\title{Gets the perpendicular distance between point and the line connecting left.point and right.point.}
\usage{
GetPerpDist(point, left.point, right.point)
}
\arguments{
\item{point}{The point in the middle. Distance will be calculated between this point and the line.}

\item{left.point}{The left point that will constitute the left endpoind of the line}

\item{right.point}{The right point that will constitute the right endpoint of the line}
}
\value{
The perpendicular distance
}
\description{
This is a helper function
}
\details{
PROVIDE CITATION FOR EQUATION
}

