% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{GetTimeLength}
\alias{GetTimeLength}
\title{Returns the Length of a Time Series in Seconds}
\usage{
GetTimeLength(timeseries)
}
\arguments{
\item{timeseries}{An xts timeseries}
}
\value{
The numeric length of the time series in seconds.
}
\description{
A helper method for Query, returns the length of an xts timeseries in seconds
}

