% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetWindowVariancePDF}
\alias{GetWindowVariancePDF}
\title{Plot the PDF of Variances for Windows of a Given Length in a Time Series}
\usage{
GetWindowVariancePDF(timeseries, window.length)
}
\arguments{
\item{timeseries}{The time series}

\item{window.length}{The length of the window with which to scan through the timeseries. Length is in number of points.}
}
\value{
A list containing (1) a numeric representing the variance with the highest probability
(2) The density function, which may be plotted using the curve function. (3) The density
object itself, which contains 512 estimated (variance, probability) pairs.
}
\description{
Scans through the provided time series with a window of window.length size, and records the
variances at each offset. Then finds the probability density function for said variances
and plots it.
}
\details{
Assuming d.fun <- GetWindowVariancePDFlot() the PDF function using
curve(d.fun$density.fun(x), from= min(d.fun$density.points$x), to=max(d.fun$density.points$x) )
}

