% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query A Time Series}
\usage{
query(timeseries, df, patterns)
}
\arguments{
\item{timeseries}{An xts timeseries}

\item{df}{A Distinctive Feature of the pattern that the query is searching for.
This function should take an xts object of length 1.
Must be of the exact form function(index, timeseries), where index is numeric and timeseries is
xts. This function should return true if the desired pattern at "index" is found within
"timeseries".}

\item{patterns}{A vector of functions of the same exact form as df: function(index, timeseries),
where index is numeric and timeseries is xts. Each function should return true if the desired
pattern at "index" is found within "timeseries".}
}
\description{
Query A Time Series
}

