% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_behavior.R
\name{rising}
\alias{rising}
\title{Rising Point}
\usage{
rising(timeseries, index)
}
\arguments{
\item{timeseries}{The time series that either contains the point referenced by index, or will be evaluated
as rising or not on average itself.}

\item{index}{The index to a point within the time series.}
}
\description{
Returns TRUE if the time series is rising on average. If an index is provided, then it returns true
if the current (indexed) point is less than the next point.
}

