% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_selectors.R
\name{selectSegment}
\alias{selectSegment}
\title{Select a Time Series Segment}
\usage{
selectSegment(index, timeseries, time1, time2)
}
\arguments{
\item{index}{The index of the current point in the series}

\item{timeseries}{The time series as xts}

\item{time1}{The beginning of the segment specified by the number of seconds from the current point.
Use a negative (-) number to specify that the segment begins in the past.}

\item{time2}{The end of the semgent specified by the number of seconds from the current point.
Use a negative (-) number to specify that the segment ends in the past.}
}
\value{
A subset of the xts time series containing all points that fall within the segment.
}
\description{
Specify a segment of the time series in the future using seconds from the current point
}

