% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pips.R
\name{EnumeratePerDistVector}
\alias{EnumeratePerDistVector}
\title{Enumerate the per.dist Vector with Perpendicular Distances}
\usage{
EnumeratePerDistVector(timeseries.length, is.pip, timeseries, perp.dist)
}
\description{
This is a helper method meant to make the code more readable,
and is thus highly coupled to GetPIPs
}

