% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pips.R
\name{GetPerpDist}
\alias{GetPerpDist}
\title{Get the Perpendicular Distance Between Point and the line connecting left.point and right.point.}
\usage{
GetPerpDist(point, left.point, right.point)
}
\arguments{
\item{point}{The point (xts object with length=1) in the middle. Distance will be calculated between this point and the line.}

\item{left.point}{The left point (xts object with length=1) that will constitute the left endpoind of the line}

\item{right.point}{The right point (xts object with length=1) that will constitute the right endpoint of the line}
}
\value{
The numeric perpendicular distance
}
\description{
This is a helper function for EnumeratePerDist, which in turn is a helper function for
GetPIPs. Thus, these functions are highly coupled, and exist simply to make the code more
readable. This function is made available to the user because it could be useful on its own.
}
\references{
Wikipedia. Distance from a point to a line — wikipedia, the free encyclopedia, 2016. [Online; accessed 26-July-2016].
}

