% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{Query}
\alias{Query}
\title{Query A Time Series Using a Sliding Window Approach}
\usage{
Query(timeseries, pattern.template, window.length = 1.2 *
  GetTimeLength(pattern.template), spearmans.rho.threshold = 0.7)
}
\arguments{
\item{timeseries}{The time series to be queries for the pattern}

\item{pattern.template}{The time series that represents a template of the pattern being searched for}

\item{window.length}{The length of the sliding window, which the pattern.template will be matched against.
Defaults to 1.2 times the length of the template pattern.}
}
\description{
Queries a given time series using a sliding window and Spearman Ranking Correlation Coefficient for
similarity assessment between each window and a given pattern.
}

