% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_behavior.R
\name{rising}
\alias{rising}
\title{Rising Point}
\usage{
rising(timeseries, index)
}
\arguments{
\item{timeseries}{The time series that either contains the point referenced by index, or will be evaluated
as rising or not on average itself.}

\item{index}{The index to a point within the time series. The 3 point segment sentered on this point will be
evaluated as rising or not.}
}
\description{
Returns true if the point is rising within the context of the previous point and next point
(i.e. previous_point < this_point < next_point). For endpoints, it relies on the single
other adjacent point (i.e previous_point < this_point OR this_point < next_point).
}

